QOL Fixes
v/COMPLETE-Setup MainMenu so after clicking play if there is no character in character list take you to character creation.
v/COMPLETE-If there is a character available in CharacterSelection Scene, can we make it so it is automatically selected. Can also be first character in list if list is not empty.
v/COMPLETE-New characters should be equipped with some starting gold.
v/COMPLETE-Make equipment info more clear on which slot it is equipped to. Wasn't always sure which body part they went to so I just equipped them to see if they replaced something and then re-equipped it if took something away that was of lower stats.
v/COMPLETE-filters and button tabs for InventoryScene for Consumables, Armor, Weapons, Materials and buttons in the InventoryScene/ShopScene/StashScene so people can tab through them.
v/COMPLETE-Debug log in battle scene isnt showing player and enemy secondary stats like atk power crit rate toughness dodge and spellward.
v/COMPLETE-Combat Log is not showing enemy skills used.
v/COMPLETE-Debuffs are showing -- instead of just - in enemy and player info in combat.